requirements.txt
asgiref==3.8.1
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.1
whitenoise==6.9.0
